fix(langserver): complete go.lsp.dev jsonrpc2/protocol migration
Port ColumnMapper to a local span helper, rewire jsonrpc2 (Conn.Go/Done, NewStream), and regenerate the Server/Client stubs for the new interfaces.

+++ rest/api.go
@@ -67,8 +67,8 @@ func (d *lspData) position() (protocol.Position, error) {
 		return protocol.Position{}, errors.New("positionChar is not specified")
 	}
 	return protocol.Position{
-		Line:      *d.PositionLine,
-		Character: *d.PositionChar,
+		Line:      uint32(*d.PositionLine),
+		Character: uint32(*d.PositionChar),
 	}, nil
 }
